Removed explicit dependency from litehtml in the header
