+++ imhtml.hpp
@@ -2,9 +2,7 @@
 
 #include <functional>
 #include <string>
-
-#include "litehtml.h"
-#include "litehtml/html_tag.h"
+#include <map>
 
 #include "imgui.h"
 #include "imgui_internal.h"
@@ -51,23 +49,6 @@ struct Config {
  * @param attributes The attributes of the element
  */
 typedef std::function<void(ImRect bounds, std::map<std::string, std::string> attributes)> CustomElementDrawFunction;
-
-/**
- * Custom element
- */
-class CustomElement : public litehtml::html_tag {
- private:
-  std::string tag = "";
-  std::map<std::string, std::string> attributes = {};
-
- public:
-  CustomElement(const std::shared_ptr<litehtml::document> &doc, const std::string &tag,
-                std::map<std::string, std::string> attributes)
-      : litehtml::html_tag(doc), tag(tag), attributes(attributes) {}
-
-  void draw_background(litehtml::uint_ptr hdc, int x, int y, const litehtml::position *clip,
-                       const std::shared_ptr<litehtml::render_item> &ri) override;
-};
 
 /**
  * Default file loader for loading CSS files
